refine tactic API
The declaration of the goal and any other evar is separate now.
This lets coq-refiner to be more liberal: the solution
needs not to be already refined (the real solution is the refined term,
not the tactic output).  Hence Solution = hole is a good implementation
of idtac, Solution = lam _ hole x\ hole of intro, ...