:bug: Fix non uniform parameter check in lp2inductive_entry
There was an unneeded lift in the function comparing the non uniform parameters with the first binders of the type of each constructor. When testing by printing the terms, it looks like they are identical without the lift. With the lift, the match takes the last branch and wrongly outputs an error.